Remove ansi sequence from luassert output
## Details

When running in a TTY luassert defaults to adding ansi colors when
showning table differences for a failed test:
https://github.com/nvim-lua/plenary.nvim/blob/master/lua/luassert/formatters/init.lua#L255

When looking at these through the `neotest` panel everything looks
great. However diagnostics do not appear to have any ansi parsing and
instead print out the raw sequence.

The parameter that defines the color can be modified and effectively
disabled by setting it to "none".