Fix vim.cmd wrapping
Previously, the method used to wrap `vim.cmd` meant that the stub ended up
being a simple function. This would result in errors if tests/code under test
used `vim.cmd` as a table (for example, to call `vim.cmd.edit()`), which has
been an option since Neovim 0.7+.

To fix this issue, we can re-add the `setmetatable` call so that we're still
able to use these newer APIs while maintaining the wrapped behavior.